ipsec: add support for egress strict mode
Generalize the config options for Wireguard's Strict Mode, so that it can
also be enabled when using IPsec for transparent encryption.
