+++ .github/actions/e2e/ipsec.yaml
@@ -92,3 +92,16 @@
   key-one: 'rfc4106-gcm-aes'
   key-two: 'rfc4106-gcm-aes'
   skip-upgrade: 'true'
+- name: 'ipsec-10'
+  kernel: '6.6'
+  kube-proxy: 'none'
+  kpr: 'true'
+  devices: '{eth0,eth1}'
+  secondary-network: 'true'
+  tunnel: 'disabled'
+  encryption: 'ipsec'
+  encryption-strict-mode: 'true'
+  key-one: 'cbc-aes-sha256'
+  key-two: 'cbc-aes-sha256'
+  # no strict mode on v1.18
+  skip-upgrade: 'true'
