+++ bpf/Makefile
@@ -85,7 +85,7 @@ HOST_OPTIONS = $(LXC_OPTIONS) \
 	-DENABLE_IPV4:-DENABLE_IPV6:-DENCAP_IFINDEX:-DPOLICY_VERDICT_NOTIFY:-DENABLE_NODEPORT:-DENABLE_NODEPORT_ACCELERATION:-DENABLE_DSR:-DENABLE_DSR_HYBRID:-DENABLE_PREFILTER:-DENABLE_HOST_FIREWALL:-DENABLE_ICMP_RULE:-DENABLE_SRV6:-DENABLE_SRV6_SRH_ENCAP:-DENABLE_SCTP:-DENABLE_VTEP:-DENABLE_IPSEC: \
 
 ifndef MAX_HOST_OPTIONS
-MAX_HOST_OPTIONS = $(MAX_BASE_OPTIONS) -DENCAP_IFINDEX=1 -DTUNNEL_MODE=1 -DENABLE_IPSEC=1
+MAX_HOST_OPTIONS = $(MAX_BASE_OPTIONS) -DENCAP_IFINDEX=1 -DTUNNEL_MODE=1 -DENABLE_IPSEC=1 -DENCRYPTION_STRICT_MODE=1
 MAX_HOST_OPTIONS += -DENABLE_WIREGUARD=1 -DENCRYPTION_STRICT_MODE=1
 endif
 
@@ -93,7 +93,7 @@ WIREGUARD_OPTIONS = $(HOST_OPTIONS)
 
 ifndef MAX_WIREGUARD_OPTIONS
 MAX_WIREGUARD_OPTIONS = $(MAX_BASE_OPTIONS) -DENCAP_IFINDEX=1 -DTUNNEL_MODE=1 -DENABLE_WIREGUARD=1
-MAX_WIREGUARD_OPTIONS += -DENABLE_WIREGUARD=1 -DENCRYPTION_STRICT_MODE=1
+MAX_WIREGUARD_OPTIONS += -DENABLE_WIREGUARD=1
 endif
 
 XDP_OPTIONS = $(LB_OPTIONS) \
