+++ bpf/complexity-tests/510/bpf_host/3.txt
@@ -26,5 +26,6 @@
 -DENABLE_VTEP=1
 -DENABLE_SCTP=1
 -DENABLE_IPSEC=1
+-DENCRYPTION_STRICT_MODE=1
 -DEVENTS_MAP_RATE_LIMIT=1000
 -DENABLE_LOCAL_REDIRECT_POLICY=1
