+++ bpf/node_config.h
@@ -190,15 +190,13 @@
 # define LB_SELECTION		LB_SELECTION_RANDOM
 #endif
 
-#ifdef ENABLE_WIREGUARD
-# ifdef ENCRYPTION_STRICT_MODE
+#ifdef ENCRYPTION_STRICT_MODE
 #  ifndef STRICT_IPV4_NET
 #   define STRICT_IPV4_NET	0
 #  endif
 #  ifndef STRICT_IPV4_NET_SIZE
 #   define STRICT_IPV4_NET_SIZE	8
 #  endif
-# endif
 #endif
 
 #ifdef ENABLE_VTEP
