+++ bpf/tests/encrypt_host_ipsec_strict.c
@@ -1,0 +1,7 @@
+// SPDX-License-Identifier: (GPL-2.0-only OR BSD-2-Clause)
+/* Copyright Authors of Cilium */
+
+#define ENABLE_IPSEC		1
+#define ENCRYPTION_STRICT_MODE	1
+
+#include "encrypt_host.h"
