+++ cilium-cli/connectivity/builder/strict_mode_encryption.go
@@ -7,6 +7,7 @@ import (
 	"github.com/cilium/cilium/cilium-cli/connectivity/check"
 	"github.com/cilium/cilium/cilium-cli/connectivity/tests"
 	"github.com/cilium/cilium/cilium-cli/utils/features"
+	"github.com/cilium/cilium/pkg/versioncheck"
 )
 
 type strictModeEncryption struct{}
@@ -33,11 +34,18 @@ func (t strictModeEncryption) build(ct *check.ConnectivityTest, _ map[string]str
 		WithCiliumVersion(">=1.18.0").
 		WithFeatureRequirements(
 			features.RequireEnabled(features.EncryptionStrictMode),
-			// Strict mode is only supported with WireGuard
-			features.RequireMode(features.EncryptionPod, "wireguard"),
+			features.RequireEnabled(features.EncryptionPod),
 			// Strict mode always allows host-to-host tunnel traffic
 			features.RequireDisabled(features.Tunnel),
 		).
+		WithCondition(func() bool {
+			if ok, _ := ct.Features.MatchRequirements(features.RequireMode(features.EncryptionPod, "ipsec")); ok {
+				if !versioncheck.MustCompile(">=1.19.0")(ct.CiliumVersion) {
+					return false
+				}
+			}
+			return true
+		}).
 		WithScenarios(tests.PodToPodMissingIPCacheV2()).
 		WithExpectations(func(_ *check.Action) (egress, ingress check.Result) {
 			return check.ResultEgressUnencryptedDrop, check.ResultEgressUnencryptedDrop
